requirements.txt
azure-core==1.30.2
azure-cosmos==4.7.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==2.0.12
click==8.1.7
Flask==3.0.3
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.26.0
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.19
Werkzeug==3.0.3
zipp==3.19.2
